cod3r/banco_dados/requirements.txt
flake8==3.9.2
mccabe==0.6.1
mysql-connector==2.2.9
pycodestyle==2.7.0
pyflakes==2.3.1